Implemented enhancements to have variables and settings now drive base platform functionality.
